[CC] Moving old GUI variables
